refactor: renames storageActions to offersActions - etc
